(PC-28533)[PRO] feat: Move publishing logic from summary step to the preview step.
