+++ pro/src/screens/CollectiveOfferSummaryCreation/CollectiveOfferSummaryCreation.tsx
@@ -1,94 +1,36 @@
-import { useState } from 'react'
-import { useNavigate, useParams } from 'react-router-dom'
+import { useParams } from 'react-router-dom'
 
 import {
   GetCollectiveOfferResponseModel,
   GetCollectiveOfferTemplateResponseModel,
-  GetOffererResponseModel,
 } from 'apiClient/v1'
 import ActionsBarSticky from 'components/ActionsBarSticky'
 import CollectiveOfferSummary from 'components/CollectiveOfferSummary'
-import { computeURLCollectiveOfferId } from 'core/OfferEducational/utils/computeURLCollectiveOfferId'
-import useActiveFeature from 'hooks/useActiveFeature'
-import useNotification from 'hooks/useNotification'
-import { RedirectToBankAccountDialog } from 'screens/Offers/RedirectToBankAccountDialog'
-import { Button, ButtonLink } from 'ui-kit'
+import { ButtonLink } from 'ui-kit'
 import { ButtonVariant } from 'ui-kit/Button/types'
 
-import publishCollectiveOfferAdapter from './adapters/publishCollectiveOfferAdapter'
-import publishCollectiveOfferTemplateAdapter from './adapters/publishCollectiveOfferTemplateAdapter'
 import styles from './CollectiveOfferSummaryCreation.module.scss'
 
 interface CollectiveOfferSummaryCreationProps {
   offer:
     | GetCollectiveOfferTemplateResponseModel
     | GetCollectiveOfferResponseModel
-  setOffer: (
-    offer:
-      | GetCollectiveOfferResponseModel
-      | GetCollectiveOfferTemplateResponseModel
-  ) => void
-  offerer: GetOffererResponseModel | undefined
 }
 
 const CollectiveOfferSummaryCreation = ({
   offer,
-  setOffer,
-  offerer,
 }: CollectiveOfferSummaryCreationProps) => {
-  const [displayRedirectDialog, setDisplayRedirectDialog] = useState(false)
-
-  const notify = useNotification()
-  const navigate = useNavigate()
-
   const { requete: requestId } = useParams()
-
-  const isNewBankDetailsJourneyEnabled = useActiveFeature(
-    'WIP_ENABLE_NEW_BANK_DETAILS_JOURNEY'
-  )
 
   const nextRedirectionUrl = offer.isTemplate
     ? `/offre/${offer.id}/collectif/vitrine/creation/apercu`
-    : `/offre/${computeURLCollectiveOfferId(
-        offer.id,
-        offer.isTemplate
-      )}/collectif/confirmation`
+    : `/offre/${offer.id}/collectif/creation/apercu`
 
   const backRedirectionUrl = offer.isTemplate
     ? `/offre/collectif/vitrine/${offer.id}/creation`
     : `/offre/${offer.id}/collectif/visibilite${
         requestId ? `?requete=${requestId}` : ''
       }`
-
-  const publishOffer = async () => {
-    if (offer.isTemplate) {
-      const response = await publishCollectiveOfferTemplateAdapter(offer.id)
-      if (!response.isOk) {
-        return notify.error(response.message)
-      }
-      setOffer(response.payload)
-      return navigate(nextRedirectionUrl)
-    }
-
-    const response = await publishCollectiveOfferAdapter(offer.id)
-    if (!response.isOk) {
-      return notify.error(response.message)
-    }
-    setOffer(response.payload)
-    const shouldDisplayRedirectDialog =
-      isNewBankDetailsJourneyEnabled &&
-      response.payload.isNonFreeOffer &&
-      offerer &&
-      !offerer.hasNonFreeOffer &&
-      !offerer.hasValidBankAccount &&
-      !offerer.hasPendingBankAccount
-
-    if (shouldDisplayRedirectDialog) {
-      setDisplayRedirectDialog(true)
-    } else {
-      navigate(nextRedirectionUrl)
-    }
-  }
 
   return (
     <div className={styles['summary']}>
@@ -111,8 +53,6 @@ const CollectiveOfferSummaryCreation = ({
           >
             Étape précédente
           </ButtonLink>
-        </ActionsBarSticky.Left>
-        {offer.isTemplate ? (
           <ButtonLink
             variant={ButtonVariant.PRIMARY}
             link={{
@@ -122,17 +62,8 @@ const CollectiveOfferSummaryCreation = ({
           >
             Étape suivante
           </ButtonLink>
-        ) : (
-          <Button onClick={publishOffer}>Publier l’offre</Button>
-        )}
+        </ActionsBarSticky.Left>
       </ActionsBarSticky>
-      {displayRedirectDialog && offerer?.id && (
-        <RedirectToBankAccountDialog
-          cancelRedirectUrl={nextRedirectionUrl}
-          offerId={offerer.id}
-          venueId={offer.venue.id}
-        />
-      )}
     </div>
   )
 }
